To provide more realistic centipawn estimate
The current function for converting winrate to centipawns is too conservative. Leela rarely shows winrates above 95% in play and this is currently converted to about 2.65 centipawns. This has no functional impact on play but is too low a level to allow sensible adjudication criteria to be set in GUIs for tournaments. (Note that for GUI adjudication purposes typically both engines have to agree that the score exceeds a set limit, for a given number of moves.) 

This tan function was suggested by jhellis3 and typically yields centipawn scores of 600-900 when Leela is almost certainly winning or losing, broadly consistent with the score of the position by other engines.